controller: Refactor graphql schema into pkg
